fix 7e2e6581cf0c530453b8cb883cc956b86f3bfd10

+++ web/ASC.Web.Api/Api/Settings/WebhooksController.cs
@@ -342,7 +342,7 @@ public class WebhooksController(
     /// <path>api/2.0/settings/webhook/triggers</path>
     /// <collection>list</collection>
     [Tags("Settings / Webhooks")]
-    [SwaggerResponse(200, "List of triggers for a webhook", typeof(Dictionary<string, int>))]
+    [SwaggerResponse(200, "List of triggers for a webhook", typeof(Dictionary<string, long>))]
     [HttpGet("webhook/triggers")]
     public Dictionary<string, long> GetWebhookTriggers()
     {
